+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### ✨ Improved
+
+* GORT will fail to initialise if the Overwatcher is running. This can be overridden by passing `override_overwatcher=True` to the `Gort` constructor or `--override-overwatcher` in the CLI.
+
+
 ## 1.0.1 - October 30, 2024
 
 ### ✨ Improved