+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * GORT will fail to initialise if the Overwatcher is running. This can be overridden by passing `override_overwatcher=True` to the `Gort` constructor or `--override-overwatcher` in the CLI.
 * Rearranged the Overwatcher helpers and make the pre- and post-observing scripts recipes.
 * Added a framework for run daily tasks and a pre-observing task.
+* Add option to disable the overwatcher to the shutdown recipe.
 
 
 ## 1.0.1 - October 30, 2024